Novo commit

+++ src/main/java/com/teamscollab/manager/core/service/TeamServiceImpl.java
@@ -5,6 +5,7 @@ import com.teamscollab.manager.core.repository.TeamRepository;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
+import java.util.ArrayList;
 import java.util.List;
 
 @Service
@@ -36,12 +37,56 @@ public class TeamServiceImpl implements TeamService {
     }
 
     @Override
-    public void deleteTeam(Integer id) {
+    public boolean deleteTeam(Integer id) {
+
+        if(!teamRepository.findById(id).isPresent())
+            return false;
+
         teamRepository.deleteById(id);
+        return true;
     }
 
     @Override
     public List<TeamEntity> listAllTeams() {
         return teamRepository.findAll();
+    }
+
+    @Override
+    public TeamEntity listById (Integer id) {
+        return teamRepository.getOne(id);
+    }
+
+    /**
+     * Returns two teams, containing the one with most wins and with most losses.
+     *
+     * @return List<TeamEntity> -> The team that has most wins, the team that has the most losses;
+     */
+    @Override
+    public List<TeamEntity> mostWinnerMostLoser() {
+
+        List<TeamEntity> teamsToCompare = listAllTeams();
+
+        // Condição para caso não existam times
+        if(teamsToCompare.isEmpty())
+            return null;
+
+        TeamEntity winningTeam = new TeamEntity();
+        TeamEntity losingTeam = new TeamEntity();
+
+        for(TeamEntity comparingTeam : teamsToCompare) {
+            if (comparingTeam.getWins() > winningTeam.getWins()) {
+                winningTeam = comparingTeam;
+            }
+
+            if (comparingTeam.getLosses() > losingTeam.getLosses()) {
+                losingTeam = comparingTeam;
+            }
+        }
+
+        List<TeamEntity> winnerLoser = new ArrayList<>();
+        winnerLoser.add(winningTeam);
+        winnerLoser.add(losingTeam);
+
+        return winnerLoser;
     }
 }
